Fix PHP 7.4 and 8.0 CI
