+++ tests/PHPStan/Analyser/nsrt/closure-passed-to-type-nested.php
@@ -149,19 +149,3 @@ acceptNestedGeneric([
 		assertType("'x'", $value);
 	}],
 ]);
-
-/**
- * @template T
- * @param array<T, \Closure(T): void> $callbacks
- */
-function acceptKeyedGenericWithUnpack(array $callbacks): void {}
-
-/** @param array<'z', \Closure('z'): void> $more */
-function withUnpack(array $more): void {
-	acceptKeyedGenericWithUnpack([
-		...$more,
-		'foo' => function ($value): void {
-			assertType("'foo'|'z'", $value);
-		},
-	]);
-};
